requirements.txt
avalanche_lib==0.6.0
easydict==1.10
joblib==1.1.1
matplotlib==3.7.1
memory_profiler==0.60.0
numpy==1.22.1
pandas==2.0.3
PyYAML==6.0.2
scikit_learn==1.0.2
scipy==1.7.3
tabulate==0.8.10
torch==1.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tqdm==4.64.1
